+++ src/mastra/index.ts
@@ -2,7 +2,7 @@ import { Mastra } from '@mastra/core/mastra';
 import { WinstonMastraLogger } from '../utils/logger/winstonMastraLogger';
 import { evergreenAgent } from './agents/evergreenAgent';
 import { parsleyOrchestrator } from './networks';
-import { historyWorkflow, versionWorkflow } from './workflows';
+import { historyWorkflow, versionWorkflow } from './workflows/evergreen';
 
 export const mastra: Mastra = new Mastra({
   workflows: {
