+++ src/mastra/workflows/evergreen/versionWorkflow.ts
@@ -1,6 +1,9 @@
 import { createWorkflow, createStep } from '@mastra/core';
 import { z } from 'zod';
-import { taskToolAdapter, versionToolAdapter } from '../tools/workflowAdapters';
+import {
+  taskToolAdapter,
+  versionToolAdapter,
+} from '../../tools/workflowAdapters';
 
 const workflowInputSchema = z.object({
   taskId: z.string(),
